requirements.txt
python-twitter
python-dotenv
lxml
beautifulsoup4
aiohttp
slackclient
uvloop